+++ CHANGELOG.md
@@ -4,12 +4,14 @@
 
 (Add entries below in the order of server, console, cli, docs, others)
 
+- server: add index on the `event_id` column of the `hdb_cron_event_invocation_logs` table
 - server: fix GraphQL type for remote relationship field (close #7284)
 - server: support EdDSA algorithm and key type for JWT
 - server: fix GraphQL type for single-row returning functions (close #7109)
 - console: add support for creation of indexes for Postgres data sources
 - console: allow same named queries and unnamed queries on allowlist file upload
 - console: support computed fields in permission builder
+
 
 ## v2.0.6